✨ pages 내부 로직 widget으로 분리
