+++ src/pages/postManager/ui/PostManager.tsx
@@ -1,31 +1,21 @@
 import { useEffect } from 'react';
 
 import { Plus } from 'lucide-react';
-import { useLocation, useNavigate } from 'react-router-dom';
 
 import {
-  useCommentActions,
-  useComments,
   useLimit,
   useLoading,
   usePostActions,
-  usePosts,
-  useSelectedPost,
   useSelectedTag,
   useSkip,
   useSortBy,
   useSortOrder,
   useTagActions,
   useDialogActions,
-  useShowPostDetailDialog,
-  useNewComment,
 } from '@/shared/model/store';
 
 import { Card, CardContent, CardHeader, CardTitle } from '@/shared/ui';
 import { Button } from '@/shared/ui';
-
-import { PostDetail } from './PostDetail';
-import { PostTable } from './PostTable';
 
 import { AddPostDialog } from '@/widgets/addPostDialog';
 import { EditPostDialog } from '@/widgets/editPostDialog';
@@ -39,16 +29,12 @@ import { useQueryParams } from '@/shared/hooks/useQueryParams';
 import { SearchPost } from '@/feature/searchPost';
 import { fetchPostsByTag, fetchPostsWithUsers } from '@/entities/post';
 import { EditCommentDialog } from '@/widgets/editCommentDialog';
+import { PostTable } from '@/widgets/postTable';
+import { PostDetail } from '@/widgets/postDetail';
 
 export const PostsManager = () => {
   useQueryParams();
 
-  const navigate = useNavigate();
-  const location = useLocation();
-  const queryParams = new URLSearchParams(location.search);
-
-  const posts = usePosts();
-  const selectedPost = useSelectedPost();
   const loading = useLoading();
 
   const { setPosts, setTotal, setLoading } = usePostActions();
@@ -60,21 +46,9 @@ export const PostsManager = () => {
 
   const selectedTag = useSelectedTag();
 
-  const { setTags, setSelectedTag } = useTagActions();
+  const { setTags } = useTagActions();
 
-  const comments = useComments();
-  const newComment = useNewComment();
-
-  const { setComments, setSelectedComment, setNewComment } = useCommentActions();
-
-  const showPostDetailDialog = useShowPostDetailDialog();
-
-  const {
-    setShowAddDialog,
-    setShowAddCommentDialog,
-    setShowEditCommentDialog,
-    setShowPostDetailDialog,
-  } = useDialogActions();
+  const { setShowAddDialog } = useDialogActions();
 
   const fetchPosts = async () => {
     setLoading(true);
@@ -102,21 +76,6 @@ export const PostsManager = () => {
     }
   };
 
-  // 댓글 삭제
-  const deleteComment = async (id, postId) => {
-    try {
-      await fetch(`/api/comments/${id}`, {
-        method: 'DELETE',
-      });
-      setComments({
-        ...comments,
-        [postId]: comments[postId].filter((comment) => comment.id !== id),
-      });
-    } catch (error) {
-      console.error('댓글 삭제 오류:', error);
-    }
-  };
-
   useEffect(() => {
     fetchTags();
   }, []);
@@ -129,25 +88,6 @@ export const PostsManager = () => {
     }
   }, [skip, limit, sortBy, sortOrder, selectedTag]);
 
-  const handleClickTag = (tag: string) => {
-    setSelectedTag(tag);
-
-    if (selectedTag) queryParams.set('tag', selectedTag);
-
-    navigate(`?${queryParams.toString()}`);
-  };
-
-  const handleClickAddComment = (postId: string) => {
-    setNewComment({ ...newComment, postId });
-    setShowAddCommentDialog(true);
-  };
-
-  const handleClickEditComment = (comment: any) => {
-    setSelectedComment(comment);
-    setShowEditCommentDialog(true);
-  };
-
-  // PostManagerPage return
   return (
     <main className='flex-grow container mx-auto px-4 py-8'>
       <Card className='w-full max-w-6xl mx-auto'>
@@ -160,9 +100,9 @@ export const PostsManager = () => {
             </Button>
           </CardTitle>
         </CardHeader>
+
         <CardContent>
           <div className='flex flex-col gap-4'>
-            {/* 검색 및 필터 컨트롤 widget*/}
             <div className='flex gap-4'>
               <SearchPost />
               <SortByTag />
@@ -170,16 +110,7 @@ export const PostsManager = () => {
               <SortByOrder />
             </div>
 
-            {/* 게시물 테이블 */}
-            {loading ? (
-              <div className='flex justify-center p-4'>로딩 중...</div>
-            ) : (
-              <PostTable
-                posts={posts}
-                selectedTag={selectedTag}
-                onClickTag={handleClickTag}
-              />
-            )}
+            {loading ? <div className='flex justify-center p-4'>로딩 중...</div> : <PostTable />}
 
             <Pagination />
           </div>
@@ -191,16 +122,7 @@ export const PostsManager = () => {
         <AddCommentDialog />
         <EditCommentDialog />
 
-        {/* 게시물 상세 보기 대화상자 */}
-        <PostDetail
-          comments={comments}
-          isOpen={showPostDetailDialog}
-          onClickOpenChange={setShowPostDetailDialog}
-          selectedPost={selectedPost}
-          onClickAdd={handleClickAddComment}
-          onClickEdit={handleClickEditComment}
-          onClickDelete={deleteComment}
-        />
+        <PostDetail />
 
         <UserInfo />
       </Card>
